<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Follow the instructions for writing a release note: https://git.k8s.io/community/contributors/guide/release-notes.md
3. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

**What type of PR is this?**
/kind cleanup

> Uncomment only one ` /kind <>` line, hit enter to put that in a new line, and remove leading whitespaces from that line:
>
> /kind api-change
> /kind bug
> /kind cleanup
> /kind design
> /kind documentation
> /kind failing-test
> /kind feature
> /kind flake

**What this PR does / why we need it**:

This PR does a series cleanups and refactors about protosanitizer:
- Update the referenced CSI spec package to 1.10
- Remove all the reference to old github.com/golang/protobuf package
- Refactor the protosanitizer to use the new protoreflect package
- Sync the Makefile for csitest.proto with csi spec repo

About the output change of protosanitizer:
- All kinds of fields now supported: including those in array, in map, or in oneof.
- Enum now outputs their name instead of number, I think this is more suitable for logging purpose.
- The nesting level of `oneof` is reduced by one.
Example diff of the output (manually formatted for diff):
```diff
--- old.json
+++ new.json
@@ -12,22 +12,18 @@
     "secrets": "***stripped***",
     "volume_capabilities": [
         {
-            "AccessType": {
-                "Mount": {
-                    "fs_type": "ext4",
-                    "mount_flags": [
-                        "flag1",
-                        "flag2",
-                        "flag3"
-                    ]
-                }
-            },
             "access_mode": {
-                "mode": 5
+                "mode": "MULTI_NODE_MULTI_WRITER"
+            },
+            "mount": {
+                "fs_type": "ext4",
+                "mount_flags": [
+                    "flag1",
+                    "flag2",
+                    "flag3"
+                ]
             }
         }
     ],
-    "volume_content_source": {
-        "Type": null
-    }
+    "volume_content_source": {}
 }
```

Also added a benchmark to justify the new implementation:
Old:
```
goos: darwin
goarch: arm64
pkg: github.com/kubernetes-csi/csi-lib-utils/protosanitizer
cpu: Apple M2 Pro
BenchmarkStrip             98264             11801 ns/op            7488 B/op        149 allocs/op
BenchmarkStripLarge           22          49160320 ns/op        39795573 B/op     620137 allocs/op
```

New:
```
BenchmarkStrip            141199              7608 ns/op            4816 B/op         89 allocs/op
BenchmarkStripLarge          122          12642834 ns/op         9955752 B/op     140009 allocs/op
```

The new implementation is significantly faster, and allocates less memory

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

**Special notes for your reviewer**:

protosanitizer now imports csi spec directly, instead of copy over the `e_CsiSecret`. Since this module already requires the csi spec module, this should not break anyone.

Support for CSI spec v0.3 is dropped, since the `csi.pb.go` file in that version does not implement the new protoreflect interface. And again, since this module already requires the csi spec module v1, users cannot use this module with v0.3 even without this PR. So, this is not a new breakage.

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note
The output format of protosanitizer is slightly refined.
```