@huww98 2024-12-14
@jsafrane Can you also take a look at this one? It should speedup several times for cases like https://github.com/kubernetes-csi/external-provisioner/pull/1310 when the log level is high.