@huww98 2024-12-17
https://github.com/golang/protobuf/issues/1655 We may get official redact support in the future. But even after that is ready, I think we should still keep ours, because that one will be prototext based, not json.